Move media functions to new file
